refactor(RC-09 02): Create ShowPosts Trait
